src/motor_control_node/requirements.txt

---

### ✅ `requirements.txt`

```txt
rclpy
std_msgs
pyserial
